requirements.txt
streamlit==1.27.2
numpy==1.24.4
Pillow==9.5.0
opencv-python-headless==4.8.1.78
tensorflow==2.12.0
